✨ feat: config 페이지 작성
